refactor(security): update imports to new api module
