+++ security/src/main/java/io/github/tgkit/security/antispam/AntiSpamInterceptor.java
@@ -15,10 +15,10 @@
  */
 package io.github.tgkit.security.antispam;
 
-import io.github.tgkit.internal.BotRequest;
+import io.github.tgkit.api.BotRequest;
 import io.github.tgkit.internal.BotResponse;
 import io.github.tgkit.internal.config.BotGlobalConfig;
-import io.github.tgkit.internal.interceptor.BotInterceptor;
+import io.github.tgkit.api.interceptor.BotInterceptor;
 import io.github.tgkit.security.captcha.CaptchaProvider;
 import io.github.tgkit.security.event.SecurityBotEvent;
 import io.github.tgkit.security.ratelimit.RateLimiter;
