+++ security/src/main/java/io/github/tgkit/security/captcha/provider/RecaptchaWebProvider.java
@@ -16,10 +16,10 @@
 package io.github.tgkit.security.captcha.provider;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.github.tgkit.internal.BotRequest;
+import io.github.tgkit.api.BotRequest;
 import io.github.tgkit.internal.config.BotGlobalConfig;
-import io.github.tgkit.internal.dsl.Button;
-import io.github.tgkit.internal.exception.BotApiException;
+import io.github.tgkit.api.dsl.Button;
+import io.github.tgkit.api.exception.BotApiException;
 import io.github.tgkit.security.captcha.CaptchaProvider;
 import io.github.tgkit.security.secret.SecretStore;
 import java.net.URI;
